fix: show total txs, fix return for all, pass progress
